+++ src/Plutarch/Extra/AssetClass.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE UndecidableInstances #-}
 
 module Plutarch.Extra.AssetClass (